Improve function return types and error handling
Now in case of errors, all functions return NULL (if return type
is a pointer) or -1 (if return type is int). Also, functions that
always succeed or their return value is always ignored now don't
have a return type (simply void). Also, errors are propagated
through all functions -- this is a first step to remove longjmp.